+++ scripts/cpu/state-idle.gd
@@ -1,0 +1,7 @@
+var target
+
+func calculatePriority(distance, enemy_position, balloons):
+	return 1
+
+func getInput(input, character):
+	return input